+++ appligator/src/appligator/airflow/gen_workflow_dag.py
@@ -3,6 +3,7 @@
 #  https://opensource.org/license/apache-2-0.
 
 from appligator.airflow.ir import workflow_to_ir
+from appligator.airflow.models import ResourceRequirements
 from appligator.airflow.renderer import AirflowRenderer
 from procodile.workflow import WorkflowStepRegistry
 
@@ -12,6 +13,7 @@ def gen_workflow_dag(
     registry: WorkflowStepRegistry,
     image: str,
     env_from_secrets: list[str] | None = None,
+    resources: ResourceRequirements | None = None,
 ) -> str:
     """Generates a fully-formed Airflow DAG Python file."""
 
@@ -22,7 +24,13 @@ def gen_workflow_dag(
         raise ValueError("Image name is required to generate dag.")
 
     # operator-agnostic intermediate representation
-    ir = workflow_to_ir(registry, dag_id, image_name=image, env_from_secrets=env_from_secrets)
+    ir = workflow_to_ir(
+        registry,
+        dag_id,
+        image_name=image,
+        env_from_secrets=env_from_secrets,
+        resources=resources,
+    )
 
     dag_code = AirflowRenderer().render(ir)